backend/requirements.txt
fastapi
uvicorn[standard]
tensorflow>=2.9.0
numpy
Pillow
python-multipart